requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
aiosmtplib
python-multipart
passlib
email-validator
jinja2
python-dotenv
bcrypt
psycopg2-binary